initial form refactor
- rename classNames so its prefixed with "form" for clarity
- extract all the answer creators to a different file
