+++ src/components/FormCreator/FormCreator.scss
@@ -4,7 +4,7 @@
 QUESTION & ANSWER UNIT
 ==========================
 */
-.voyage-application-QA {
+.form-QA {
     display: block;
     padding-top: 20px;
 }
@@ -13,7 +13,7 @@ QUESTION & ANSWER UNIT
 QUESTION
 ==========================
 */
-.voyage-application-question {
+.form-question {
     font-size: 16px;
     color: $dark-grey;
     display: block;
@@ -24,7 +24,7 @@ QUESTION
 QUESTION SUBTEXT
 ==========================
 */
-.voyage-application-subtext {
+.form-subtext {
     font-size: 12px;
     color: $grey;
     display: inline-block;
@@ -36,13 +36,13 @@ QUESTION SUBTEXT
 ANSWER FONT
 ==========================
 */
-.voyage-application-answer {
+.form-answer {
     font-size: 13.5px;
     color: $dark-grey;
     margin-left: -10px;
     padding-left: 10px;
 }
-.voyage-application-checkbox-answer {
+.form-checkbox-answer {
     font-size: 13.5px;
     color: $dark-grey;
     width: fit-content;
@@ -54,7 +54,7 @@ ANSWER FONT
 INPUT & TEXTAREA
 ==========================
 */
-.voyage-application-input {
+.form-input {
     display: block;
     width: 100%;
     font-size: 14px;
@@ -64,7 +64,7 @@ INPUT & TEXTAREA
     border: 1px solid $light-grey;
 }
 
-.voyage-application-text-area {
+.form-text-area {
     display: block;
     width: 100%;
     font-size: 14px;
@@ -95,7 +95,7 @@ ERROR TEXT
 DROPDOWN
 ==========================
 */
-.voyage-application-dropdown {
+.form-dropdown {
     width: 100%;
     border: 1px solid $light-grey;
     padding: 10px;
@@ -103,7 +103,7 @@ DROPDOWN
     font-size: 13.5px;
     margin-top: 5px;
 }
-.voyage-application-dropdown > option {
+.form-dropdown > option {
     width: 418px !important;
 }
 /*
