+++ src/components/FormCreator/answerCreators.js
@@ -1,363 +1,57 @@
-import * as React from 'react';
-
-class CheckboxAnswerCreator extends React.Component {
-  render() {
-    const { answer, questionId, index, onFormChange, state } = this.props;
-    return (
-      <div key={'checkbox-answer_' + questionId + '_' + index} className="checkbox-container">
-        <label className="voyage-application-checkbox-answer" htmlFor={questionId + '_' + index}>
-          {answer}
-          <input
-            type="checkbox"
-            name={questionId}
-            value={answer}
-            id={questionId + '_' + index}
-            checked={state[questionId].has(answer)}
-            onChange={e => onFormChange(e)}
-          />
-          <span className="checkmark" />
-        </label>
-      </div>
-    )
-  }
-}
-
-class RadioAnswerCreator extends React.Component {
-  render() {
-    const { answer, questionId, index, onFormChange, state } = this.props;
-    return (
-      <div key={'radio-answer_' + questionId + '_' + index} className="radio-container">
-        <label className="voyage-application-answer" htmlFor={questionId + '_' + index}>
-          {answer}
-          <input
-            type="radio"
-            name={questionId}
-            id={questionId + '_' + index}
-            value={answer}
-            checked={state[questionId] === answer}
-            onChange={e => onFormChange(e)}
-          />
-          <span className="radio-checkmark" />
-        </label>
-      </div>
-    )
-  }
-}
-class RadioWithInputAnswerCreator extends React.Component {
-  render() {
-    const { answer, questionId, index, onFormChange, state } = this.props;
-    return (
-      <div key={'radio-answer_' + questionId + '_' + index} className="radio-container">
-        <label className="voyage-application-answer" htmlFor={questionId + '_' + index}>
-          {answer}
-          <input
-            className="voyage-application-radio"
-            type="radio"
-            name={questionId}
-            id={questionId + '_' + index}
-            value={answer}
-            checked={state[questionId] === answer}
-            onChange={e => onFormChange(e)}
-          />
-          <span className="radio-checkmark" />
-        </label>
-      </div>
-    )
-  }
-  // render() {
-  //     const { answer, questionId, index, onFormChange, state } = this.props;
-  //     return (
-  //         <div key={'radio-answer_' + questionId + '_' + index} className="radio-container">
-  //             <input type="text"
-  //                 name={questionId}
-  //                 value={'Other: ' + state[questionId]}
-  //                 onChange={e => onFormChange(e)}
-  //                 className="voyage-application-input"
-  //             />
-  //             <label className="voyage-application-answer" htmlFor={questionId + '_' + index}>
-  //                 {answer}
-  //                 <input
-  //                     className="voyage-application-radio"
-  //                     type="radio"
-  //                     name={questionId}
-  //                     id={questionId + '_' + index}
-  //                     value={answer}
-  //                     checked={state[questionId].includes(answer)}
-  //                     onChange={e => onFormChange(e)}
-  //                 />
-  //                 <span className="radio-checkmark" />
-  //             </label>
-  //         </div>
-  //     )
-  // }
-}
-
-function answerCreator_checkbox(data, onFormChange, state) {
-  return (
-    data.answers.map((answer, index) => {
-      return <CheckboxAnswerCreator key={data.id + '_' + index} answer={answer} questionId={data.id} index={index} onFormChange={onFormChange} state={state} />
-    })
-  )
-}
-
-
-function answerCreator_radio(data, onFormChange, state) {
-  return (
-    data.answers.map((answer, index) => {
-      if (answer === 'Other') {
-        return <RadioWithInputAnswerCreator key={data.id + '_' + index} answer={answer} questionId={data.id} index={index} onFormChange={onFormChange} state={state} />
-      }
-      return <RadioAnswerCreator key={data.id + '_' + index} answer={answer} questionId={data.id} index={index} onFormChange={onFormChange} state={state} />
-    })
-  )
-}
-class SpecialRadioBadgeAnswerCreator extends React.Component {
-  render() {
-    const badges = [
-      {
-        img: require('../../assets/tier badges-01.png'),
-        className: 'badge',
-        btnClassName: 'badge--1'
-      },
-      {
-        img: require('../../assets/tier badges-02.png'),
-        className: 'badge',
-        btnClassName: 'badge--2'
-      },
-      {
-        img: require('../../assets/tier badges-03.png'),
-        className: 'badge',
-        btnClassName: 'badge--3'
-      }
-    ]
-    const { answer, questionId, index, onFormChange, state } = this.props;
-    return (
-      <div className={badges[index].btnClassName + '-container'}>
-        <div key={'radio-answer_' + questionId + '_' + index} className="radio-container">
-          <label className="voyage-application-answer" htmlFor={questionId + '_' + index}>
-            <div className={badges[index].btnClassName}>
-              <img className={badges[index].className} src={badges[index].img} alt={'badge-' + index} />
-            </div>
-            <div className="badge-title">{answer.title}</div>
-            <div className="badge-subtext">{answer.subtext}</div>
-            <input
-              className="voyage-application-radio special-badge-input"
-              type="radio"
-              name={questionId}
-              id={questionId + '_' + index}
-              value={answer.title}
-              checked={state[questionId] === answer.title}
-              onChange={e => onFormChange(e)}
-            />
-            <span className="radio-checkmark--badge" />
-          </label>
-        </div>
-      </div>
-    )
-  }
-}
-function answerCreator_radio_special_badge(data, onFormChange, state) {
-  return (
-    <div className="badge-container">
-      {data.answers.map((answer, index) => {
-        return <SpecialRadioBadgeAnswerCreator key={data.id + '_' + index} answer={answer} questionId={data.id} index={index} onFormChange={onFormChange} state={state} />
-      })
-      }
-    </div>
-  )
-}
-
-function answerCreator_input(data, onFormChange, state) {
-  return (
-    <input type="text"
-      name={data.id}
-      value={state[data.id]}
-      onChange={e => onFormChange(e)}
-      className="voyage-application-input"
-      minLength={10}
-    />
-  )
-}
-
-function answerCreator_textarea(data, onFormChange, state) {
-  return (
-    <textarea type="text"
-      name={data.id}
-      value={state[data.id]}
-      onChange={e => onFormChange(e)}
-      className="voyage-application-text-area"
-      minLength={10}
-    />
-  )
-}
-
-function answerCreator_dropdown(data, onFormChange, state) {
-  return (
-    <select
-      className="voyage-application-dropdown"
-      value={state[data.id]}
-      name={data.id}
-      type="dropdown"
-      onChange={e => onFormChange(e)}
-      multiple={false}
-    >
-      {data.answers.map((answer, index) => {
-        return (
-          <option className="voyage-application-answer" value={answer} key={'dropdown_' + data.id + '_' + index}>{answer}</option>
-        )
-      })}
-    </select>
-  )
-}
-
-function answerCreator_dropdown_multiple(data, onFormChange, state) {
-  const renderMultiple = () => {
-    return data.answers.map((answer, index) => {
-      return <CheckboxAnswerCreator key={data.id + '_' + index} answer={answer} questionId={data.id} index={index} onFormChange={onFormChange} state={state} />
-    }
-    );
-  };
-
-  const filter = (filterId, elemByName) => {
-    var filter, inputOptions;
-    filter = document.getElementById(filterId).value.toUpperCase();
-    inputOptions = document.getElementsByName(elemByName);
-    for (var i = 0; i < inputOptions.length; i++) {
-      if (inputOptions[i].value.toUpperCase().indexOf(filter) !== -1) {
-        inputOptions[i].style.display = '';
-      } else {
-        inputOptions[i].style.display = 'none';
-      }
-    }
-  }
-
-  const inputBoxFilter = (e) => {
-    e.persist();
-    filter("dropdownSearch_" + data.id, data.id);
-  }
-
-  return (
-    <React.Fragment>
-      <button
-        className={"filter-dropdown-btn-" + data.id}
-      >
-        {data.placeholder}
-        <div id={"dropdown_multiple-" + data.id} className="filter-dropdown-hide filter-dropdown-box">
-          <input
-            className="search-input-box"
-            type="text"
-            placeholder="Search / Add Teckstack"
-            id={"dropdownSearch_" + data.id}
-            onKeyUp={e => inputBoxFilter(e)}
-          />
-          {renderMultiple()}
-        </div>
-      </button>
-    </React.Fragment>
-  )
-}
-
-function answerCreator_checkbox_2_column(data, onFormChange, state) {
-  let firstHalf = [];
-  let secondHalf = [];
-  for (var i = 0; i < (data.answers).length; i++) {
-    if (i < (data.answers).length / 2) {
-      firstHalf.push(data.answers[i]);
-    } else {
-      secondHalf.push(data.answers[i]);
-    }
-  }
-
-  let idx = 0;
-  return (
-    <div className="checkbox-2-column-container">
-      <div className="checkbox-column-1">
-        {firstHalf.map((answer, index) => {
-          const questionIdx = idx++;
-          return <CheckboxAnswerCreator key={data.id + '_' + questionIdx} answer={answer} questionId={data.id} index={questionIdx} onFormChange={onFormChange} state={state} />
-        })}
-      </div>
-      <div className="checkbox-column-2">
-        {secondHalf.map((answer, index) => {
-          const questionIdx = idx++;
-          return <CheckboxAnswerCreator key={data.id + '_' + questionIdx} answer={answer} questionId={data.id} index={questionIdx} onFormChange={onFormChange} state={state} />
-        })}
-      </div>
-    </div>
-  )
-}
-
-const FormButtom = ({ color, onClick, data, id }) => {
-  return (
-    <input
-      className={"btn-3 " + color}
-      type="button"
-      value={data.answer}
-      onClick={e => onClick(e)}
-      name={id}
-    />
-  );
-}
-
-function answerCreator_btn_3_options(data, onFormChange, state) {
-  return (
-    <div className="btn-container--3">
-      {data.answers.map((answer, index) => {
-        return (
-          <FormButtom
-            id={data.id}
-            key={index}
-            color={state[data.id] === answer.value ? 'btn-3--' + answer.value : ''}
-            onClick={onFormChange}
-            data={answer}
-          />
-        )
-      })}
-    </div>
-  )
-}
+import React from 'react';
+import { 
+  answerCreator_checkbox,
+  answerCreator_radio,
+  answerCreator_radio_special_badge,
+  answerCreator_input,
+  answerCreator_textarea,
+  answerCreator_dropdown,
+  answerCreator_dropdown_multiple,
+  answerCreator_checkbox_2_column,
+  answerCreator_btn_3_options
+} from './AnswerTypes/AnswerTypes';
 
 export function renderQAs(applicationData, onFormChange, state) {
   return applicationData.map((setOfQuestionAnswer) => {
     let answerComponent;
     switch (setOfQuestionAnswer.type) {
       case 'checkbox':
-        answerComponent = answerCreator_checkbox(setOfQuestionAnswer, onFormChange, state);
+        answerComponent = answerCreator_checkbox;
         break;
       case 'input':
-        answerComponent = answerCreator_input(setOfQuestionAnswer, onFormChange, state);
+        answerComponent = answerCreator_input;
         break;
       case 'radio':
-        answerComponent = answerCreator_radio(setOfQuestionAnswer, onFormChange, state);
+        answerComponent = answerCreator_radio;
         break;
       case 'dropdown':
-        answerComponent = answerCreator_dropdown(setOfQuestionAnswer, onFormChange, state)
+        answerComponent = answerCreator_dropdown;
         break;
       case 'dropdown-multiple':
-        answerComponent = answerCreator_dropdown_multiple(setOfQuestionAnswer, onFormChange, state)
+        answerComponent = answerCreator_dropdown_multiple;
         break;
       case 'textarea':
-        answerComponent = answerCreator_textarea(setOfQuestionAnswer, onFormChange, state)
+        answerComponent = answerCreator_textarea;
         break;
       case 'checkbox-2-column':
-        answerComponent = answerCreator_checkbox_2_column(setOfQuestionAnswer, onFormChange, state)
+        answerComponent = answerCreator_checkbox_2_column;
         break;
       case 'radio-special-badge':
-        answerComponent = answerCreator_radio_special_badge(setOfQuestionAnswer, onFormChange, state)
+        answerComponent = answerCreator_radio_special_badge;
         break;
       case 'button-3-colors':
-        answerComponent = answerCreator_btn_3_options(setOfQuestionAnswer, onFormChange, state)
+        answerComponent = answerCreator_btn_3_options;
         break;
       default:
         break;
     }
     return (
-      <div key={'question_' + setOfQuestionAnswer.id} className="voyage-application-QA">
-        <label className="voyage-application-question">
+      <div key={'question_' + setOfQuestionAnswer.id} className="form-QA">
+        <label className="form-question">
           {setOfQuestionAnswer.question}
         </label>
-        {setOfQuestionAnswer.subtext ? <div className="voyage-application-subtext">{setOfQuestionAnswer.subtext}</div> : null}
-        {answerComponent}
+        {setOfQuestionAnswer.subtext ? <div className="form-subtext">{setOfQuestionAnswer.subtext}</div> : null}
+        {answerComponent(setOfQuestionAnswer, onFormChange, state)}
       </div>
     )
   })
